requirements.txt
streamlit
torch
torchvision
pillow
numpy
scikit-learn
sqlalchemy
requests